Fix Errors from W3C Nu Html Checker
Error 1 to 3  on report related to duplicate attribute at Lines
24, 25, 26.  Removed the second class attribute in each href
and incorporated them into one class for each link.
Error 4 Center Element is obsolete - inserted a div element
to replace the center element  with a class attribute of
video-container which is used in style.css to apply center
alignment to the video. Error 5 Bad value for attribute controls
on video removed the value for controls.  Error 6 type not allowed
on element video so removed the type attribute from the video.
Error 7 no space between attributes in audio element so removed
the comma a the end of src in the audio element to add space.
Error 8, 9 also relate to the above comma at point no 7.
Error 10 bad value for attribute controls on element audio.
Here removed the value given to controls.  Error 11 an img
element mus have an alt attribute - added an alt attribute
to the image tag.  Error 12  Trailing slash on void elements
- removed  trailing slash at the end of the of the image line of
code.