Add Responsiveness For Page 2
Complete responsiveness for page 2 of website.  In
specifications.html changed div class name container-animation to
a more meaningfull explaination like container-images and also
changed conatiner-animation on line 85 to container-images2
to make it more meaningfull.  Changed the name of the class at
line 126 to footerstyling instead of foot-style-social-media so
as to avoid duplication of code.