Add mobile responsiveness for page 3
Add mobile responsiveness for contact-us.html with
media queries at the end of the css file. On specifications page
change heading from Specifications to Specifications Page so that
the user can identify what page they are  currently visiting.
Fix radio buttons in contact us page so that the user can
only select one time for a viewing. Remove style properties
from line 77, 85 and 90 as it was interfering with the mobile view.
Add iframe-styling to line 111 on contact-us to allow for making
google maps responsive on mobile device.