Add favicon and fix positing of google image
Add favicon icon to root and also added link to index.html.
In Dev tools there was errors re ids in the contact page so
in contact.html added id of "name" to the Name section
as it had no id.
Changed id of area_one under message to message.
Added id of submit to the Submit button as there was no id.
Add styling to center google image on contact page to css file.