Fix h2 heading on contact us page
The h2 header in the footer element on the contact us page was not
displaying "Find us on the following meida platforms and continue the
conversation" renamed div class name at line 120 to match index and
specifications page renamed to footersytling.  In style.css
renamed foot-style-social-media elements to footer styling and
aligned h2 to the left of the screen.