Add Html W3C Validation to README
Contact.html add blank option to property of interest at line 83.
Fix return to home button at bottom of screen as it was not
accepting an "a" attribute inside a button so just changed it to a
"a" attribute. Index.html changed the return to home button as
per the contact.html page. Specifications.html removed blank space
at line 58 and fixed issued with return to home button as per
home.html above.  Style.css added styling for return to home button.