Add link from contact page to success page.
Add link for success page to submit button in contact us
page.  Style the submit button from the contact us page in the
style.css file.  Style the thank you message on the success page with
code for same on the style.css file and remove the return to home
button as it is not required.